requirements.txt

# tkinter # usually installed with python
freetype-py
numpy
Pillow
packaging
